+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle case where the redirect target is bad / unsupported (#332 and #356)
 - Fixed WARC files handling order to follow creation order (#366)
 - Remove subsequent slashes in URLs, both in Python and JS (#365)
+- Ignore non HTTP(S) WARC records (#351)
 
 ## [2.0.3] - 2024-07-24